requirements.txt
streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4   
pandas==2.1.0
openpyxl==3.1.2